refactor : move common response dto
