+++ src/main/java/coz/weavon/exception/handler/GlobalExceptionHandler.java
@@ -2,8 +2,8 @@ package coz.weavon.exception.handler;
 
 import coz.weavon.exception.model.BusinessException;
 import coz.weavon.exception.model.ClientException;
-import coz.weavon.io.model.ErrorResponse;
-import coz.weavon.io.model.RestResponse;
+import coz.weavon.io.model.response.ErrorResponse;
+import coz.weavon.io.model.response.RestResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
