+++ src/main/java/coz/weavon/io/model/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package coz.weavon.io.model;
+package coz.weavon.io.model.response;
 
 import lombok.Builder;
 import lombok.Getter;
