+++ src/main/java/coz/weavon/io/model/response/ResponseType.java
@@ -1,4 +1,4 @@
-package coz.weavon.io.model;
+package coz.weavon.io.model.response;
 
 public enum ResponseType {
     SUCCESS,
